Move GOMAXPROCS adjustment to beatcmd
This is something that should run just once for
the lifetime of the process. I'm intending to
move the reload logic out of beater and into
beatcmd (or a subpackage).
